refactor: move ArchUnit test to root module for full classpath scanning
Move ProductNameArchTest from module-product to src/test so it scans
all classes across the entire project, not just module-product.
