+++ build.gradle.kts
@@ -35,6 +35,7 @@ dependencies {
     testImplementation("io.cucumber:cucumber-java:7.20.1")
     testImplementation("io.cucumber:cucumber-spring:7.20.1")
     testImplementation("io.cucumber:cucumber-junit-platform-engine:7.20.1")
+    testImplementation("com.tngtech.archunit:archunit-junit5:1.3.0")
     testImplementation("org.junit.platform:junit-platform-suite-api")
     testRuntimeOnly("org.junit.platform:junit-platform-launcher")
     testRuntimeOnly("org.junit.platform:junit-platform-suite-engine")
