+++ module-product/build.gradle.kts
@@ -10,13 +10,3 @@ java {
         languageVersion = JavaLanguageVersion.of(21)
     }
 }
-
-dependencies {
-    testImplementation("com.tngtech.archunit:archunit-junit5:1.3.0")
-    testImplementation("org.junit.jupiter:junit-jupiter:5.11.0")
-    testRuntimeOnly("org.junit.platform:junit-platform-launcher")
-}
-
-tasks.withType<Test> {
-    useJUnitPlatform()
-}
